models/requirements.txt
gradio==4.41.0
ffmpeg-python==0.2.0
transformers==4.44.0 
openai-whisper==20231117
rembg==2.0.58
